+++ infra/postgres-init.sql
@@ -1,1 +1,2 @@
 CREATE DATABASE applicant;
+CREATE DATABASE investigator;